🔒️ Moved AWS signature generation to api
