+++ .env.example
@@ -2,10 +2,10 @@ BYTESHARE_VERSION=0.3.2
 ENVIRONMENT=dev
 NEXT_PUBLIC_API_BASE_URL=http://localhost:8000
 NEXT_PUBLIC_APPWRITE_PROJECT_ID=
-NEXT_PUBLIC_AWS_REGION=us-east-2
-NEXT_PUBLIC_AWS_ACCESS_KEY_ID=
-NEXT_PUBLIC_AWS_SECRET_ACCESS_KEY=
-NEXT_PUBLIC_ROLE_ARN=
+AWS_REGION=us-east-2
+AWS_ACCESS_KEY_ID=
+AWS_SECRET_ACCESS_KEY=
+ROLE_ARN=
 NEXT_PUBLIC_API_KEY=thisisthekey
 WEB_BASE_URL=http://localhost:3000
 RESEND_API_KEY=
