+++ ui/.env.example
@@ -3,7 +3,7 @@ NEXT_PUBLIC_API_KEY=thisisthekey
 NEXT_PUBLIC_APP_URL=http://localhost:3000
 NEXT_PUBLIC_APPWRITE_URL=https://cloud.appwrite.io/v1
 NEXT_PUBLIC_APPWRITE_PROJECT_ID=
-NEXT_PUBLIC_AWS_REGION=us-east-2
-NEXT_PUBLIC_AWS_ACCESS_KEY_ID=
-NEXT_PUBLIC_AWS_SECRET_ACCESS_KEY=
-NEXT_PUBLIC_ROLE_ARN=
+AWS_REGION=us-east-2
+AWS_ACCESS_KEY_ID=
+AWS_SECRET_ACCESS_KEY=
+ROLE_ARN=
